feat(auth): add ratings and favorites backend services

+++ MovieScanner/Entities/Rating.cs
@@ -7,7 +7,7 @@
         public User User { get; set; } = null!;
         public int MediaId { get; set; }
         public string MediaType { get; set; } = string.Empty; // movie / tv
-        public int Score { get; set; } // 1 à 5
+        public decimal Score { get; set; } // 1 à 5
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
